Functional rootfinding code

+++ rootfinding/rootfinder.h
@@ -4,8 +4,8 @@
 #include <random>
 
 double bisector(
-    double f(double, int, double, double, double),
-    const int dim, const double lambda, const double beta,
+    double f(double, double, double, double, double),
+    const double dim, const double lambda, const double beta,
     const double beta_critical, const std::random_device::result_type entropy);
 
 #endif
